refactor: :recycle: update import path for RegisterBook component and add RegisterBook.vue file
